Add --make-executable
Now we can load a parser and dump the core image to an executable.
This means that instead of

$ cat to-be-parsed | ./bp parser-name --run

we can do

$ ./bp parser-name --make-executable
$ cat to-be-parsed | ./parser-name

I've chosen to make it dump the core image instead of creating a
script with #!/usr/bin/sbcl --script at the top, mainly because I
don't want to worry about loading sbcl at runtime. The downside of
this is that the executable file is gigantic. (39M)!

Might want to consider having
    --make-shell-script-executable
and
    --dump-core-image-to-executable
(obviously with better names)
so that we have the option of having a lower file size, at the cost
of a higher runtime.